# SelectVariants can now write VCF to cloud