--- a/src/main/java/org/broadinstitute/hellbender/engine/GATKTool.java
+++ b/src/main/java/org/broadinstitute/hellbender/engine/GATKTool.java
@@ -772,14 +772,28 @@ public abstract class GATKTool extends CommandLineProgram {
 
     /**
      * Creates a VariantContextWriter whose outputFile type is determined by
-     * the outFile's extension, using the best available sequence dictionary for
+     * the outPathName's extension, using the best available sequence dictionary for
      * this tool, and default index, leniency and md5 generation settings.
+     *
+     * Deprecated, use {@link #createVCFWriter(Path)} instead.
      *
      * @param outFile output File for this writer. May not be null.
      * @returns VariantContextWriter must be closed by the caller
      */
     public VariantContextWriter createVCFWriter(final File outFile) {
-        Utils.nonNull(outFile);
+        return createVCFWriter(outFile == null ? null : outFile.toPath());
+    }
+
+    /**
+     * Creates a VariantContextWriter whose outputFile type is determined by
+     * outPath's extension, using the best available sequence dictionary for
+     * this tool, and default index, leniency and md5 generation settings.
+     *
+     * @param outPath output Path for this writer. May not be null.
+     * @returns VariantContextWriter must be closed by the caller
+     */
+    public VariantContextWriter createVCFWriter(final Path outPath) {
+        Utils.nonNull(outPath);
 
         final SAMSequenceDictionary sequenceDictionary = getBestAvailableSequenceDictionary();
 
@@ -802,7 +816,7 @@ public abstract class GATKTool extends CommandLineProgram {
         }
 
         return GATKVariantContextUtils.createVCFWriter(
-                outFile,
+                outPath,
                 sequenceDictionary,
                 createOutputVariantMD5,
                 options.toArray(new Options[options.size()]));