… src/test/java/org/broadinstitute/hellbender/tools/walkers/variantutils/SelectVariantsIntegrationTest.java
… src/test/java/org/broadinstitute/hellbender/tools/walkers/vqsr/VariantGaussianMixtureModelUnitTest.java
… src/test/java/org/broadinstitute/hellbender/utils/variant/GATKVariantContextUtilsUnitTest.java
… src/test/java/org/broadinstitute/hellbender/utils/variant/writers/GVCFWriterUnitTest.java