NIO output support for SelectVariants. Tested like so:

```
$ ./gatk SelectVariants \
  --variant dbsnp_138.b37.excluding_sites_after_129.vcf \
  --select-random-fraction 0.01 \
  --output gs://mybucket/variants.vcf
$ gsutil ls -lh gs://mybucket/*.vcf
 23.38 MiB  2018-10-30T23:58:12Z  gs://mybucket/variants.vcf
```

Includes the required changes under the hood, plus test updates.

This change also gives NIO support to **HaplotypeCaller**, so it can write its VCF to cloud storage.

Also, fixes #2128.