@jean-philippe-martin 2018-12-01
I've added a test to write via NIO, but instead of writing to an actual bucket I'm using the jimfs in-memory filesystem NIO provider. This still checks that the VCF code can use NIO paths, but it runs a lot faster.