@jean-philippe-martin 2019-01-18
rebased