@cmnbroad 2019-01-18
@jean-philippe-martin Thanks for adding the additional test, but by "integration test", I meant something that exercises an actual tool (which is why I mentioned SelectVariants) with a non-default provider, not another unit test that uses GCS. I suggested SelectVariants since I thought it would be easy:

> all the previous comments have been addressed with the exception of adding a SelectVariants integration test. It should be pretty easy to clone an existing case and change it use a non-default nio provider

I think this last test is redundant with the one you already added. My apologies if that was confusing.