@jean-philippe-martin 2019-01-18
I misunderstood, sorry. Does "non-default provider" mean in this case "run SelectVariants on GCS" ?