@jean-philippe-martin 2019-01-18
I've added a new end-to-end test for SelectVariants that writes to GCS.

Sadly, the IntegrationTestSpec class uses Files throughout, so it wasn't possible to do this simply without first completely refactoring IntegrationTestSpec (which should probably be its own pull request). 

Doing this refactoring would have the advantage that changing existing end-to-end tests from local to GCS would be trivial.

For now instead I went with an ad-hoc approach. It works, and the test passes.