@jean-philippe-martin 2019-01-18
Also, I agree that this PR should make HaplotypeCaller also able to write to GCS, since it uses GATKVariantContextUtils and that one gains GCS support with this pull request.