@jean-philippe-martin 2019-01-23
@cmnbroad Thank you for pointing out those build failures and even digging down to the apparent cause! I investigated and the issue wasn't inability to decompress gzip files (no test output was compressed), but that XReadLines trims the lines by default. The "expected" files have an extra tab at the end of some lines (the CHROM line for example) that this was picking up.

What I've done is updated XReadLines so it can take Paths as input, so we get good matching behavior without having to duplicate code. While I was at it I also exposed XReadLines' ability to strip out comments, so assertEqualTextFiles didn't need to re-implement it anymore.

Assuming Travis passes, this should be ready to review. I have the feeling we're getting close!