@cmnbroad 2019-02-12
Totally get it now; with these changes, `assertEqualTextFiles` is getting `LinkedList`s from `XReadLines`, rather than the `ArrayList`s which it used to get from the trailing `stream.collect` calls. Since it then iterates thorough the two lists zip style, it used indexed access, and the serial `get` calls are super slow on even moderately sized (100k) linked lists. Should be an easy fix.