@jean-philippe-martin 2019-02-12
@cmnbroad thank you so much for looking into this carefully! I have pushed a fix to the issue. In addition to using an `ArrayList` in `XReadLines` (to speed up all callers), I also fixed `assertEqualTextFiles` to do a streaming comparison so it uses less memory (and might even run a bit faster).