@jean-philippe-martin 2019-02-12
Travis passed. I looked at a couple of tests, and their time is back to the pre-regression levels. Sadly we're not faster, but we're about the same. On the plus side, we can now write VCF to cloud for SelectVariants and HaplotypeCaller!

Test | time before | time while broken | time now
-- | -- | -- | --
`VariantRecalibratorIntegrationTest` | 1m29s | 25m50s | 1m43s
`BaseRecalibratorIntegrationTest` | 43s | 5m1s | 46s