@jean-philippe-martin 2019-02-25
OK here's a squashed and rebased version. Let's see how the tests go.