@jean-philippe-martin 2019-02-26
Looks good. I also re-ran `gatk SelectVariants` with a GCS input and output as above and it ran to completion.